Update pre-commit and invoke tasks, python 3.11
Minor tune-ups to invoke and pre-commit tasks.

Update the docker image with python 3.11 and latest versions of other
pythons.

Disable support for python 3.6.
