+++ .pre-commit-config.yaml
@@ -16,17 +16,28 @@ repos:
       - id: check-yaml # checks yaml files for parseable syntax.
       - id: destroyed-symlinks # detects symlinks which are changed to regular files with a content of a path which that symlink was pointing to.
       - id: detect-private-key # detects the presence of private keys.
+        exclude: tools/crx-example/getting-started\.pem
       - id: end-of-file-fixer # ensures that a file is either empty, or ends with one newline.
+        exclude: |
+          (?x)^(
+              ^tools/crx-example/.*
+          )$
+
       - id: file-contents-sorter # sorts the lines in specified files (defaults to alphabetical). you must provide list of target files as input in your .pre-commit-config.yaml file.
       - id: fix-byte-order-marker # removes utf-8 byte order marker.
       - id: mixed-line-ending # replaces or checks mixed line ending.
       - id: pretty-format-json # This hook sets a standard for formatting JSON files.
-        exclude: .vscode
+        exclude: |
+          (?x)^(
+              ^\.vscode/.*|
+              ^tools/crx-example/.*
+          )$
+
       - id: requirements-txt-fixer # sorts entries in requirements.txt.
       - id: trailing-whitespace # trims trailing whitespace.
 
   - repo: https://github.com/psf/black
-    rev: 22.8.0
+    rev: 22.10.0
     hooks:
       - id: black
 
@@ -34,10 +45,11 @@ repos:
     rev: 5.10.1
     hooks:
       - id: isort
+        args: [--profile=black]
         files: (dtrx)|(\.(py)$)
 
   - repo: https://github.com/sirosen/check-jsonschema
-    rev: 0.18.3
+    rev: 0.19.1
     hooks:
       - id: check-github-actions
       - id: check-github-workflows
@@ -53,30 +65,50 @@ repos:
         files: \./README\.md
 
   - repo: https://github.com/pre-commit/mirrors-prettier
-    rev: v3.0.0-alpha.0
+    rev: v3.0.0-alpha.4
     hooks:
       - id: prettier
         args: [--write]
         files: \.(js|ts|jsx|tsx|css|less|html|json|markdown|md|yaml|yml)$
-        exclude: .vscode
+        exclude: |
+          (?x)^(
+              ^\.vscode/.*|
+              ^tools/crx-example/.*
+          )$
 
-  - repo: https://github.com/pryorda/dockerfilelint-precommit-hooks
-    rev: v0.1.0
+  - repo: https://github.com/hadolint/hadolint
+    rev: v2.12.1-beta
     hooks:
-      - id: dockerfilelint
+      - id: hadolint-docker
 
-  - repo: https://gitlab.com/pycqa/flake8
-    rev: 3.9.2
+  - repo: https://github.com/PyCQA/flake8
+    rev: 5.0.4
     hooks:
       - id: flake8
         additional_dependencies: [flake8-bugbear]
 
-  - repo: https://github.com/shellcheck-py/shellcheck-py
-    rev: v0.8.0.4
+  - repo: https://github.com/koalaman/shellcheck-precommit
+    rev: v0.8.0
     hooks:
       - id: shellcheck
+        files: \.sh$
 
   - repo: https://github.com/asottile/setup-cfg-fmt
-    rev: v2.0.0
+    rev: v2.2.0
     hooks:
       - id: setup-cfg-fmt
+
+  - repo: local
+    hooks:
+      # pylint needs to run from local python environment, so it can check
+      # imported packages
+      - id: pylint
+        name: pylint
+        entry: pylint
+        language: system
+        types: [python]
+        exclude: |
+          (?x)^(
+              ^tests/.*|
+              ^dtrx/dtrx.py
+          )$
