+++ .python-version
@@ -1,6 +1,7 @@
-2.7.16
+2.7.18
 3.6.15
-3.7.13
-3.8.13
-3.9.12
-3.10.4
+3.7.15
+3.8.15
+3.9.15
+3.10.8
+3.11.0
