+++ requirements.txt
@@ -5,5 +5,5 @@ docutils==0.16
 invoke==1.7.1
 pre-commit==2.20.0
 pyyaml==5.4.1
-tox==3.15.2
+tox==3.26.0
 tox-pyenv==1.1.0
