+++ tasks/utils.py
@@ -1,0 +1,8 @@
+"""
+Invoke utilities shared module.
+"""
+
+
+from pathlib import Path
+
+ROOT_DIR = Path(__file__).absolute().parent.parent
