+++ tox.ini
@@ -7,7 +7,8 @@
 requires = [ "setuptools == 40.6.3", "wheel == 0.32.3"]
 
 [tox]
-envlist = py27, py36, py37, py38, py39, py3.10
+# disabling py36 for now, something broken
+envlist = py27, py37, py38, py39, py3.10, py3.11
 
 [testenv]
 whitelist_externals =
